change grade level and corrector summary points to float

+++ src/Corrector/Rest.php
@@ -218,7 +218,7 @@ class Rest extends Base\BaseRest
                     if ($corrector->getKey() == $CurrentCorrectorKey) {
                         $summary = new CorrectionSummary(
                             isset($data['text']) ? (string) $data['text'] : null,
-                            isset($data['points']) ? (int) $data['points'] : null,
+                            isset($data['points']) ? (float) $data['points'] : null,
                             isset($data['grade_key']) ? (string) $data['grade_key'] : null,
                             isset($data['last_change']) ? (int) $data['last_change'] : time(),
                             isset($data['is_authorized']) ? (bool) $data['is_authorized'] : null,
